backend/ml-worker/requirements.txt
torch==2.1.1
torchvision==0.16.1
numpy==1.24.4
pillow==10.1.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
pika==1.3.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
structlog==23.2.0
prometheus-client==0.19.0
aiofiles==23.2.0
httpx==0.25.2
redis==5.0.1
albumentations==1.3.1
matplotlib==3.8.2
seaborn==0.13.0